# gateway/gateway-controller/api/management-openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Controller Management API
  description: |
    REST API for managing API configurations in the WSO2 API Platform Gateway.
  version: 1.0.0
  contact:
    name: WSO2 API Platform Team
servers:
  - url: http://localhost:9090/api/management/v1
    description: Local development
  - url: http://gateway-controller:9090/api/management/v1
    description: Docker/Kubernetes deployment
security:
  - basicAuth: []
paths:
  /mcp:
    post:
      summary: Model Context Protocol endpoint
      description: Single MCP endpoint implementing the Streamable HTTP transport of the Model Context Protocol (2026-07-28).
      operationId: handleMcp
      x-basicauth-roles: [admin, developer, consumer]
      tags:
        - MCP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: A single JSON-RPC 2.0 request object.
      responses:
        "200":
          description: |
            JSON-RPC response. Returned as a single JSON object, or as an SSE
            stream carrying request-scoped notifications followed by the response.
          content:
            application/json:
              schema:
                type: object
            text/event-stream:
              schema:
                type: string
        "400":
          description: Malformed JSON-RPC message, header/body mismatch, or unsupported protocol version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: |
            Missing or invalid credentials. Carries a `WWW-Authenticate: Bearer`
            challenge with `resource_metadata` and the scopes to request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: |
            The credential is valid but lacks a scope required for the requested
            tool. Carries a `WWW-Authenticate: Bearer` challenge with
            `error="insufficient_scope"` and the required scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: The MCP endpoint is not enabled on this gateway
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body exceeds the configured maximum
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rest-apis:
    post:
      summary: Create a new RestAPI
      description: Add a new RestAPI to the Gateway.
      operationId: createRestAPI
      x-basicauth-roles: [admin, developer]
      tags:
        - Rest API Management
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/RestAPIRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/RestAPIRequest"
      responses:
        "201":
          description: RestAPI created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestAPI"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - API with same name and version already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List all RestAPIs
      description: List RestAPIs registered in the Gateway, optionally filtered by name, version, context, or status.
      operationId: listRestAPIs
      x-basicauth-roles: [admin, developer]
      tags:
        - Rest API Management
      parameters:
        - name: displayName
          in: query
          required: false
          description: Filter by API display name
          schema:
            type: string
          example: Reading List API
        - name: version
          in: query
          required: false
          description: Filter by API version
          schema:
            type: string
          example: v1.0
        - name: context
          in: query
          required: false
          description: Filter by API context/path
          schema:
            type: string
          example: /reading-list
        - name: status
          in: query
          required: false
          description: Filter by deployment status
          schema:
            type: string
            enum: [ deployed, undeployed ]
          example: undeployed
      responses:
        "200":
          description: List of RestAPIs
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  count:
                    type: integer
                    example: 5
                  apis:
                    type: array
                    items:
                      $ref: "#/components/schemas/RestAPI"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /certificates:
    get:
      summary: List all custom certificates
      description: |
        Retrieve all custom TLS certificates currently loaded in the certificate store.
        These certificates are used for verifying HTTPS upstream connections.
      operationId: listCertificates
      x-basicauth-roles: [admin, developer]
      tags:
        - Certificate Management
      responses:
        "200":
          description: List of certificates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateListResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Upload a new certificate
      description: Upload a new TLS certificate (PEM format) to the Gateway. The certificate is loaded dynamically without restarting the Gateway.
      operationId: uploadCertificate
      x-basicauth-roles: [admin]
      tags:
        - Certificate Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateUploadRequest"
            example:
              name: my-custom-ca
              certificate: |
                -----BEGIN CERTIFICATE-----
                MIIDXTCCAkWgAwIBAgIJAKL0UG+mRKtjMA0GCSqGSIb3DQEBCwUAMEUxCzAJBgNV
                ...
                -----END CERTIFICATE-----
      responses:
        "201":
          description: Certificate uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateResponse"
        "400":
          description: Invalid certificate format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    RestAPIRequest:
      type: object
      required:
        - apiVersion
        - metadata
        - kind
        - spec
      properties:
        apiVersion:
          type: string
          description: API specification version
          example: gateway.api-platform.wso2.com/v1
          enum:
            - gateway.api-platform.wso2.com/v1
        kind:
          type: string
          description: API type
          example: RestApi
          enum:
            - RestApi
        metadata:
          $ref: "#/components/schemas/Metadata"
        spec:
          $ref: '#/components/schemas/APIConfigData'
      example:
        apiVersion: gateway.api-platform.wso2.com/v1
        kind: RestApi
        metadata:
          name: reading-list-api-v1.0
        spec:
          displayName: Reading-List-API
          version: v1.0
          context: /reading-list/$version
          upstream:
            main:
              url: https://apis.bijira.dev/samples/reading-list-api-service/v1.0
          policies:
            - name: set-headers
              version: v1
              params:
                request:
                  headers:
                    - name: x-wso2-apip-gateway-version
                      value: v1.0.0
                response:
                  headers:
                    - name: x-environment
                      value: development
          operations:
            - method: GET
              path: /books
            - method: POST
              path: /books
            - method: GET
              path: /books/{id}
            - method: PUT
              path: /books/{id}
            - method: DELETE
              path: /books/{id}

    # Full resource including server-managed status (list/get responses).
    RestAPI:
      allOf:
        - $ref: '#/components/schemas/RestAPIRequest'
        - type: object
          properties:
            status:
              readOnly: true
              description: Server-managed lifecycle fields. Populated on responses.
              allOf:
                - $ref: '#/components/schemas/ResourceStatus'
      example:
        apiVersion: gateway.api-platform.wso2.com/v1
        kind: RestApi
        metadata:
          name: reading-list-api-v1.0
        spec:
          displayName: Reading-List-API
          version: v1.0
          context: /reading-list/$version
          upstream:
            main:
              url: https://apis.bijira.dev/samples/reading-list-api-service/v1.0
          policies:
            - name: set-headers
              version: v1
              params:
                request:
                  headers:
                    - name: x-wso2-apip-gateway-version
                      value: v1.0.0
                response:
                  headers:
                    - name: x-environment
                      value: development
          operations:
            - method: GET
              path: /books
            - method: POST
              path: /books
            - method: GET
              path: /books/{id}
            - method: PUT
              path: /books/{id}
            - method: DELETE
              path: /books/{id}
        status:
          id: reading-list-api-v1.0
          state: deployed
          createdAt: 2026-04-24T07:21:13Z
          updatedAt: 2026-04-24T07:21:13Z
          deployedAt: 2026-04-24T07:21:13Z
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          description: High-level error description
          example: Configuration validation failed
        errors:
          type: array
          description: Detailed validation errors
          items:
            $ref: "#/components/schemas/ValidationError"
...